+++ src/client/containers/app.js
@@ -38,6 +38,7 @@ const App = ({socket}) => {
         socket.emit(`room#${room_name}`, {token, command: cmd.START})
       }}>START</button>
       <textarea id='tmpScreen' rows="30" cols="50"></textarea>
+      <textarea id='tmpScreen2' rows="30" cols="50"></textarea>
       <script> {maincode(socket)}</script>
     </div>
   )
@@ -54,6 +55,8 @@ function maincode(socket){
     if (msg.command == "REFRESH_PLAYER")
       if (msg.data.socketid == socket.id)
         document.getElementById('tmpScreen').value = JSON.stringify(msg.data.screen).replace(/],/g,'\n').replace(/\[/g,'').replace(']]','')
+      else
+        document.getElementById('tmpScreen2').value = JSON.stringify(msg.data.screen).replace(/],/g,'\n').replace(/\[/g,'').replace(']]','')
   })
 
   // Ecoute les key pressed et les envoies au back
